refactor(tools): 实验性 A/B 脚本归拢到 tools/experimental
移动 affinity_ab/llm_prompt_lab/llm_persona_ab/semantic_style_audit/simulate_decision_ab 至 tools/experimental，同步更新对应测试的 import 路径。
